fix: combines functions into a component event-edit.js
