--- src/components/events/event-edit/destination-item.js
@@ -1,7 +1,0 @@
-const createDestinationItemTemplate = (city) => {
-  return (
-    `<option value="${city}"></option>`
-  );
-};
-
-export {createDestinationItemTemplate};
