--- src/components/events/event-edit/photo-tape.js
@@ -1,7 +1,0 @@
-const createTripPhotoTemplate = (src) => {
-  return (
-    `<img class="event__photo" src="${src}" alt="Event photo">`
-  );
-};
-
-export {createTripPhotoTemplate};
